third_party/blink/web_tests/fast/scrolling/scrollbars/mouse-scrolling-on-div-scrollbar.html

<!DOCTYPE>
<title>Clicking on a non-custom scrollbar should scroll its scroller</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<style>
.mouseDiv {
  width: 100px;
  height: 100px;
  overflow: scroll;
  border: 1px solid black;
  position: absolute;
  top: 100px;
  left: 100px;
}
.fast {
  will-change: transform;
}
.space {
  height: 1000px;
  width: 1000px;
}
</style>

<!-- Composited fast scroller -->
<div class="mouseDiv fast">
  <div class="space"></div>
</div>

<script>
if (window.internals)
    internals.settings.setScrollAnimatorEnabled(false);

window.onload = () => {
  const divFast = document.querySelector('.fast');
  const rectFast = divFast.getBoundingClientRect();
  const BUTTON_WIDTH = 15;
  const SCROLL_CORNER = BUTTON_WIDTH;

  promise_test (async () => {
    await waitForCompositorCommit();

    // Click on the Down arrow for rectFast.
    let x = rectFast.right - BUTTON_WIDTH / 2;
    let y = rectFast.bottom - SCROLL_CORNER - BUTTON_WIDTH / 2;
    await mouseClickOn(x, y);
    assert_equals(divFast.scrollTop, 40, "Pressing the down arrow didn't scroll.");

    // Click on the Up arrow for rectFast.
    x = rectFast.right - BUTTON_WIDTH / 2;
    y = rectFast.top + BUTTON_WIDTH / 2;
	await mouseClickOn(x, y);
    assert_equals(divFast.scrollTop, 0, "Pressing the up arrow didn't scroll.");

    // Click on the Right arrow for rectFast.
    x = rectFast.right - SCROLL_CORNER - BUTTON_WIDTH / 2;
    y = rectFast.bottom - BUTTON_WIDTH / 2;
	await mouseClickOn(x, y);
    assert_equals(divFast.scrollLeft, 40, "Pressing the right arrow didn't scroll.");

    // Click on the Left arrow for rectFast.
    x = rectFast.left + BUTTON_WIDTH / 2;
    y = rectFast.bottom - BUTTON_WIDTH / 2;
	await mouseClickOn(x, y);
    assert_equals(divFast.scrollLeft, 0, "Pressing the left arrow didn't scroll.");
  }, "Test mouse click on non-custom scrollbars.");
}
</script>
